REL-2143: Provide generic RuleSpec specification for future tests.
